fix(tests): move EntityValidationTests to active BitwardenMacOSTests target
Xcode created a new empty EntityValidationTests.swift in BitwardenMacOSTests/
instead of linking the existing one from Tests/DomainTests/Entities/.
Copy real test content to the target folder and remove the orphan file.
